refactor(punishments): remove dependency to core
