+++ punishments/src/main/kotlin/gg/airbrush/punishments/Punishments.kt
@@ -15,7 +15,7 @@ package gg.airbrush.punishments
 import cc.ekblad.toml.decode
 import cc.ekblad.toml.model.TomlException
 import cc.ekblad.toml.tomlMapper
-import gg.airbrush.core.lib.setInterval
+import gg.airbrush.sdk.lib.setInterval
 import gg.airbrush.punishments.commands.*
 import gg.airbrush.punishments.events.*
 import gg.airbrush.sdk.SDK
